feat(func): partition by for analytic functions runtime
