+++ internal/binder/function/funcs_misc.go
@@ -479,151 +479,6 @@ func registerMiscFunc() {
 		exec:  nil, // directly return in the valuer
 		val:   ValidateNoArg,
 	}
-	builtins["changed_col"] = builtinFunc{
-		fType: ast.FuncTypeScalar,
-		exec: func(ctx api.FunctionContext, args []interface{}) (interface{}, bool) {
-			ignoreNull, ok := args[0].(bool)
-			if !ok {
-				return fmt.Errorf("first arg is not a bool but got %v", args[0]), false
-			}
-			if ignoreNull && args[1] == nil {
-				return nil, true
-			}
-			lv, err := ctx.GetState("self")
-			if err != nil {
-				return err, false
-			}
-			if !reflect.DeepEqual(args[1], lv) {
-				err := ctx.PutState("self", args[1])
-				if err != nil {
-					return err, false
-				}
-				return args[1], true
-			}
-			return nil, true
-		},
-		val: func(_ api.FunctionContext, args []ast.Expr) error {
-			if err := ValidateLen(2, len(args)); err != nil {
-				return err
-			}
-			if ast.IsNumericArg(args[0]) || ast.IsTimeArg(args[0]) || ast.IsStringArg(args[0]) {
-				return ProduceErrInfo(0, "boolean")
-			}
-			return nil
-		},
-	}
-	builtins["had_changed"] = builtinFunc{
-		fType: ast.FuncTypeScalar,
-		exec: func(ctx api.FunctionContext, args []interface{}) (interface{}, bool) {
-			if len(args) <= 1 {
-				return fmt.Errorf("expect more than one arg but got %d", len(args)), false
-			}
-			ignoreNull, ok := args[0].(bool)
-			if !ok {
-				return fmt.Errorf("first arg is not a bool but got %v", args[0]), false
-			}
-			result := false
-			for i := 1; i < len(args); i++ {
-				v := args[i]
-				k := strconv.Itoa(i)
-				if ignoreNull && v == nil {
-					continue
-				}
-				lv, err := ctx.GetState(k)
-				if err != nil {
-					return fmt.Errorf("error getting state for %s: %v", k, err), false
-				}
-				if !reflect.DeepEqual(v, lv) {
-					result = true
-					err := ctx.PutState(k, v)
-					if err != nil {
-						return fmt.Errorf("error setting state for %s: %v", k, err), false
-					}
-				}
-			}
-			return result, true
-		},
-		val: func(_ api.FunctionContext, args []ast.Expr) error {
-			if len(args) <= 1 {
-				return fmt.Errorf("expect more than one arg but got %d", len(args))
-			}
-			if ast.IsNumericArg(args[0]) || ast.IsTimeArg(args[0]) || ast.IsStringArg(args[0]) {
-				return ProduceErrInfo(0, "bool")
-			}
-			return nil
-		},
-	}
-
-	builtins["lag"] = builtinFunc{
-		fType: ast.FuncTypeScalar,
-		exec: func(ctx api.FunctionContext, args []interface{}) (interface{}, bool) {
-			if len(args) != 1 && len(args) != 2 && len(args) != 3 {
-				return fmt.Errorf("expect one two or three args but got %d", len(args)), false
-			}
-			lagkey := "self"
-			v, err := ctx.GetState(lagkey)
-			if err != nil {
-				return fmt.Errorf("error getting state for %s: %v", lagkey, err), false
-			}
-			if v == nil {
-				size := 0
-				var dftVal interface{} = nil
-				if len(args) == 3 {
-					dftVal = args[2]
-				}
-				// first time call, need create state for lag
-				if len(args) == 1 {
-					size = 1
-				} else {
-					siz, ok := args[1].(int)
-					if !ok {
-						return fmt.Errorf("second arg is not a int but got %v", args[1]), false
-					}
-					size = siz
-				}
-
-				rq := newRingqueue(size)
-				rq.fill(dftVal)
-
-				rtnVal, _ := rq.fetch()
-				rq.append(args[0])
-				err := ctx.PutState(lagkey, rq)
-				if err != nil {
-					return fmt.Errorf("error setting state for %s: %v", lagkey, err), false
-				}
-				return rtnVal, true
-			} else {
-				rq, ok := v.(*ringqueue)
-				if !ok {
-					return fmt.Errorf("error getting state for %s: %v", lagkey, err), false
-				}
-				rtnVal, _ := rq.fetch()
-				rq.append(args[0])
-				err := ctx.PutState(lagkey, rq)
-				if err != nil {
-					return fmt.Errorf("error setting state for %s: %v", lagkey, err), false
-				}
-				return rtnVal, true
-			}
-		},
-		val: func(_ api.FunctionContext, args []ast.Expr) error {
-			if len(args) != 1 && len(args) != 2 && len(args) != 3 {
-				return fmt.Errorf("expect one two or three args but got %d", len(args))
-			}
-			if len(args) >= 2 {
-				if ast.IsFloatArg(args[1]) || ast.IsTimeArg(args[1]) || ast.IsBooleanArg(args[1]) || ast.IsStringArg(args[1]) || ast.IsFieldRefArg(args[1]) {
-					return ProduceErrInfo(1, "int")
-				}
-				if s, ok := args[1].(*ast.IntegerLiteral); ok {
-					if s.Val < 0 {
-						return fmt.Errorf("the index should not be a nagtive integer")
-					}
-				}
-			}
-			return nil
-		},
-	}
-
 	builtins["object_construct"] = builtinFunc{
 		fType: ast.FuncTypeScalar,
 		exec: func(ctx api.FunctionContext, args []interface{}) (interface{}, bool) {
