+++ internal/binder/function/function.go
@@ -38,6 +38,7 @@ func init() {
 	registerMathFunc()
 	registerStrFunc()
 	registerMiscFunc()
+	registerAnalyticFunc()
 	registerColsFunc()
 }
